+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Removed
 ### Fixed
 
+- `UnionConverter`: Support overriding discriminator without needing to nominate a `catchAllCase`
+
 <a name="2.1.0"></a>
 ## [2.1.0] - 2020-05-10